move components into game feature package
